* internal/develop_rel_1_5_2: (29 commits)
  Update core version to 1.5.2
  os:util:mocos: fix build error for mocOS
  Add creating attribute with related commandId function.
  easysetup: use HTTP_CONN_H handle as pointer type
  easysetup: remove unused posix testing
  Makefile: fix wrong feature handling
  iot_main: Change mqtt disconnection timing to prevent unwanted self-reconnectng
  iot_nv_data: add type casting on log printing
  logfile: modify argument of iot_os_eventgroup_wait_bits
  logfile:bugfix: add break in switch code for FLASH_WITH_RAM
  logfile: move declation code to inside of switch statement
  easysetup: fix build error with x509 easysetup feature
  test: add TCs for iot_easysetup_http.c
  TC:dump_log: modify variable declaration location
  TC:dump_log: modify to test with different log message size
  block unused switch code for IOT_CORE_LOG_FILE
  easysetup: remove unused value
  test: add TCs for iot_easysetup_http.c
  easysetup: declare internal functions as static
  Update parameter for nvs_partition_gen.py
  ...

+++ src/iot_nv_data.c
@@ -420,7 +420,7 @@ iot_error_t iot_nv_get_wifi_prov_data(struct iot_wifi_prov_data* wifi_prov)
 				ret = IOT_ERROR_NONE;
 			}
 		} else {
-			IOT_INFO("Saved AP BSSID is invalid length:%d", read_len);
+			IOT_INFO("Saved AP BSSID is invalid length:%u", (unsigned int)read_len);
 			memset(wifi_prov->mac_str, '\0', sizeof(wifi_prov->mac_str));
 			memset(wifi_prov->bssid.addr, 0, sizeof(wifi_prov->bssid.addr));
 		}
